qcacld-2.0: Fix compile issues when some macros are disabled
Fix compile issues when some macros are disabled.

CRs-Fixed: 2565286
