+++ CORE/HDD/src/wlan_hdd_p2p.c
@@ -56,6 +56,7 @@
 //Ms to Micro Sec
 #define MS_TO_MUS(x)   ((x)*1000)
 
+#ifdef WLAN_DEBUG
 static tANI_U8* hdd_getActionString(tANI_U16 MsgType)
 {
     switch (MsgType)
@@ -82,6 +83,7 @@ static tANI_U8* hdd_getActionString(tANI_U16 MsgType)
            return ("UNKNOWN");
      }
 }
+#endif
 
 #ifdef WLAN_FEATURE_P2P_DEBUG
 #define MAX_P2P_ACTION_FRAME_TYPE 9
@@ -3340,7 +3342,9 @@ void __hdd_indicate_mgmt_frame(hdd_adapter_t *pAdapter,
                 vos_mem_compare(&pbFrames[WLAN_HDD_PUBLIC_ACTION_FRAME_OFFSET+2], SIR_MAC_P2P_OUI, SIR_MAC_P2P_OUI_SIZE))
             // P2P action frames
             {
+#ifdef WLAN_DEBUG
                 u8 *macFrom = &pbFrames[WLAN_HDD_80211_FRM_DA_OFFSET+6];
+#endif
                 actionFrmType = pbFrames[WLAN_HDD_PUBLIC_ACTION_FRAME_TYPE_OFFSET];
                 hddLog(LOG1, "Rx Action Frame %u", actionFrmType);
 #ifdef WLAN_FEATURE_P2P_DEBUG
