+++ CORE/MAC/src/pe/lim/limSendManagementFrames.c
@@ -5910,7 +5910,9 @@ static void lim_tx_mgmt_frame(tpAniSirGlobal mac_ctx,
 			struct sir_mgmt_msg *mb_msg, uint32_t msg_len,
 			void *packet, uint8_t *frame)
 {
+#ifdef TRACE_RECORD
 	tpSirMacFrameCtl fc = (tpSirMacFrameCtl)mb_msg->data;
+#endif
 	eHalStatus hal_status;
 	uint8_t sme_session_id = 0;
 	tpPESession session;
