+++ CORE/SERVICES/BMI/ol_fw.h
@@ -163,7 +163,7 @@ void ol_target_failure(void *instance, A_STATUS status);
 u_int8_t ol_get_number_of_peers_supported(struct ol_softc *scn);
 
 #ifdef REMOVE_PKT_LOG
-static inline void ol_pktlog_init(void *)
+static inline void ol_pktlog_init(void *hifsc)
 {
 }
 #else
