+++ CORE/SERVICES/COMMON/pktlog_ac_fmt.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012-2013, 2015-2016 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2012-2013, 2015-2016, 2019 The Linux Foundation. All rights reserved.
  *
  * Previously licensed under the ISC license by Qualcomm Atheros, Inc.
  *
@@ -71,26 +71,6 @@ struct ath_pktlog_hdr {
     u_int32_t timestamp;
 }__ATTRIB_PACK;
 
-
-/**
- * enum pkt_type - packet type
- * @START_MONITOR: indicates parser to start packetdump parsing
- * @STOP_MONITOR: indicates parser to stop packetdump parsing
- * @TX_MGMT_PKT: TX management Packet
- * @TX_DATA_PKT: TX data Packet
- * @RX_MGMT_PKT: RX management Packet
- * @RX_DATA_PKT: RX data Packet
- *
- * This enum has packet types
- */
-enum pkt_type {
-	START_MONITOR = 1,
-	STOP_MONITOR,
-	TX_MGMT_PKT,
-	TX_DATA_PKT,
-	RX_MGMT_PKT,
-	RX_DATA_PKT,
-};
 
 #define ATH_PKTLOG_HDR_FLAGS_MASK 0xffff
 #define ATH_PKTLOG_HDR_FLAGS_SHIFT 0
@@ -423,47 +403,6 @@ enum tx_pkt_fate {
 	TX_PKT_FATE_DRV_DROP_INVALID,
 	TX_PKT_FATE_DRV_DROP_NOBUFS,
 	TX_PKT_FATE_DRV_DROP_OTHER,
-};
-
-/**
- * enum rx_pkt_fate - tx packet fate
- * @RX_PKT_FATE_SUCCESS: Valid and delivered to
- * network stack (e.g., netif_rx()).
- * @RX_PKT_FATE_FW_QUEUED: Queued within firmware,
- * but not yet sent to driver.
- * @RX_PKT_FATE_FW_DROP_FILTER: Dropped by firmware
- * due to host-programmable filters.
- * @RX_PKT_FATE_FW_DROP_INVALID: Dropped by firmware
- * as invalid. E.g. bad checksum, decrypt failed, or invalid for current state.
- * @RX_PKT_FATE_FW_DROP_NOBUFS: Dropped by firmware
- * due to lack of buffer space.
- * @RX_PKT_FATE_FW_DROP_OTHER: Dropped by firmware
- * for any other reason.
- * @RX_PKT_FATE_DRV_QUEUED: Queued within driver,
- * not yet delivered to network stack.
- * @RX_PKT_FATE_DRV_DROP_FILTER: Dropped by drive
- * r due to filter rules.
- * @RX_PKT_FATE_DRV_DROP_INVALID: Dropped by driver as invalid.
- * E.g. not permitted in current state.
- * @RX_PKT_FATE_DRV_DROP_NOBUFS: Dropped by driver
- * due to lack of buffer space.
- * @RX_PKT_FATE_DRV_DROP_OTHER: Dropped by driver for any other reason.
- *
- * This enum has packet fate types
- */
-
-enum rx_pkt_fate {
-	RX_PKT_FATE_SUCCESS,
-	RX_PKT_FATE_FW_QUEUED,
-	RX_PKT_FATE_FW_DROP_FILTER,
-	RX_PKT_FATE_FW_DROP_INVALID,
-	RX_PKT_FATE_FW_DROP_NOBUFS,
-	RX_PKT_FATE_FW_DROP_OTHER,
-	RX_PKT_FATE_DRV_QUEUED,
-	RX_PKT_FATE_DRV_DROP_FILTER,
-	RX_PKT_FATE_DRV_DROP_INVALID,
-	RX_PKT_FATE_DRV_DROP_NOBUFS,
-	RX_PKT_FATE_DRV_DROP_OTHER,
 };
 
 #endif  /* _PKTLOG_FMT_H_ */
