+++ CORE/SME/src/sme_common/sme_Api.c
@@ -12988,7 +12988,7 @@ void sme_UpdateEnableSSR(tHalHandle hHal, tANI_BOOLEAN enableSSR)
 /*
  * SME API to stringify bonding mode. (hostapd convention)
  */
-
+#ifdef WLAN_DEBUG
 static const char* sme_CBMode2String( tANI_U32 mode)
 {
    switch (mode)
@@ -13014,6 +13014,7 @@ static const char* sme_CBMode2String( tANI_U32 mode)
          return "Unknown";
    }
 }
+#endif
 
 /*
  * SME API to adjust bonding mode to regulatory, dfs nol .. etc.
