+++ CORE/UTILS/PKTLOG/include/pktlog_ac.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012-2013, 2016, 2018 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2012-2013, 2016, 2018-2019 The Linux Foundation. All rights reserved.
  *
  * Previously licensed under the ISC license by Qualcomm Atheros, Inc.
  *
@@ -140,24 +140,5 @@ int pktlog_setsize(struct ol_softc *scn, int32_t log_state);
 int pktlog_disable(struct ol_softc *scn);
 int pktlogmod_init(void *context);
 void pktlogmod_exit(void *context);
-#else /* REMOVE_PKT_LOG */
-#define ol_pktlog_attach(_scn)	({ (void)_scn; })
-#define ol_pktlog_detach(_scn)	({ (void)_scn; })
-static inline void pktlog_init(struct ol_softc *scn)
-{
-	return;
-}
-static int pktlog_enable(struct ol_softc *scn, int32_t log_state)
-{
-	return 0;
-}
-static int pktlog_setsize(struct ol_softc *scn, int32_t log_state)
-{
-	return 0;
-}
-static int pktlog_disable(struct ol_softc *scn)
-{
-	return 0;
-}
 #endif /* REMOVE_PKT_LOG */
 #endif /* _PKTLOG_AC_H_ */
